Pass database down
